# Implement batched queue mutation