--- a/impl/sql/mutationstorage/mutations.go
+++ b/impl/sql/mutationstorage/mutations.go
@@ -38,7 +38,7 @@ var (
 		DirectoryID VARCHAR(30)   NOT NULL,
 		LogID    BIGINT           NOT NULL,
 		Time     BIGINT           NOT NULL,
-		ID       BIGINT           NOT NULL,
+		Sequence BIGINT           NOT NULL,
 		Mutation BLOB             NOT NULL,
 		PRIMARY KEY(DirectoryID, LogID, Time)
 	);`,