This improves the write performance from ~10k entries/s to ~100 k/s

Before
```
BenchmarkSend-8 		10000 			101075 ns/op
```

After
```
BenchmarkSend/1-8 	   	   20000	     92531 ns/op
BenchmarkSend/2-8 	   	   10000	    104603 ns/op
BenchmarkSend/4-8 	   	   10000	    125030 ns/op
BenchmarkSend/8-8 	   	   10000	    164034 ns/op
BenchmarkSend/16-8         	    5000	    247246 ns/op
BenchmarkSend/32-8         	    3000	    406327 ns/op
BenchmarkSend/64-8         	    2000	    730439 ns/op
BenchmarkSend/128-8        	    1000	   1330174 ns/op
BenchmarkSend/256-8        	     500	   2521275 ns/op
```
![screen shot 2018-11-27 at 12 48 55](https://user-images.githubusercontent.com/198478/49082933-0414c200-f243-11e8-8c7c-2ae658d6c135.png)


Part of #1229