@pavelkalinnikov 2018-11-27
The "Before" benchmark in PR description looks to me like ~10k writes/s to me. It's in nanoseconds, right?