@gdbelvin 2018-11-27
Updated, with new, real perf numbers.